Disable edge-to-edge to make status bar visible in dark mode (#2)
A problem was identified on a Xiaomi device where the navigation bar was
displayed in white with white icons, making it virtually impossible to
see any icons shown on those two places.

Also, the status bar was displayed in a clear background color far from
what one would expect when using dark mode.

| Before | After |
|---|---|
|
![Screenshot_20250421_222800](https://github.com/user-attachments/assets/0ab25399-1e9e-4d30-b2af-7017e88a091f)
|
![Screenshot_20250421_222913](https://github.com/user-attachments/assets/7782031a-d4e0-4cc1-9c0c-ef3597ff435d)
|